ex 4.17-4.21

+++ controllers/blogs.js
@@ -3,6 +3,7 @@ const { response } = require('express')
 const { request } = require('../app')
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -10,8 +11,24 @@ blogsRouter.get('/', async (request, response) => {
   response.json(blogs.map(blog => blog.toJSON()))
 });
 
+/*
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
+*/
+
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({error: 'token misisng or invalid'})
+  }
+  const user = await User.findById(decodedToken.id)
+  /*
   let user
   if (body.userId === undefined){
     const users = await User.find({})
@@ -20,6 +37,7 @@ blogsRouter.post('/', async (request, response) => {
   else{
     user = await User.findById(body.userId)
   }
+  */
   const blog = new Blog({
     ...body,
     user: user._id
@@ -32,6 +50,17 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({error: 'token missing or invalid'})
+  }
+  const user = await User.findById(decodedToken.id)
+  const blogToRemove = await Blog.findById(request.params.id)
+  console.log("Blog user: ", blogToRemove.user.toString())
+  console.log("Current user: ", blogToRemove.user.toString())
+  if (!(blogToRemove.user.toString() === user.id.toString())){
+    return response.status(401).json({error: 'only creator of the blog is allowed to remove it'})
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
